feat: working userprofile (still missing book list)
