+++ client/app/src/main/AndroidManifest.xml
@@ -42,6 +42,7 @@
             android:parentActivityName=".Activities.RootActivity" />
         <activity android:name=".Activities.MyBookActivity" />
         <activity android:name=".Activities.BookActivity" />
+        <activity android:name=".Activities.UserProfileActivity" />
     </application>
 
 </manifest>
